+++ src/main/java/org/c4marathon/assignment/member/session/SessionMemberInfo.java
@@ -1,8 +1,10 @@
 package org.c4marathon.assignment.member.session;
 
+import java.io.Serializable;
+
 public record SessionMemberInfo(
 	long memberPk,
 	String memberId,
 	long mainAccountPk
-) {
+) implements Serializable {
 }
